CCM-1656: Add /messages endpoint with 201 response to sandbox
