+++ sandbox/.mocharc.yml
@@ -1,4 +1,4 @@
-spec: './*.spec.js'
+spec: './__test__/**/*.spec.js'
 ui: bdd
 watch-files:
   - '**/*.js'
