+++ sandbox/__test__/status.spec.js
@@ -1,0 +1,59 @@
+const request = require("supertest");
+
+const { setup } = require('./helpers');
+
+describe('status', () => {
+  let env;
+
+  const version_info = {
+    build_label:"1233-shaacdef1",
+    releaseId:"1234",
+    commitId:"acdef12341ccc"
+  };
+
+  before(function () {
+    env = process.env;
+    server = setup()
+  });
+
+  after(function () {
+      process.env = env;
+      server.close();
+  });
+
+  it("responds to /_ping", (done) => {
+    request(server)
+        .get("/_ping")
+        .expect(200, {
+            status: "pass",
+            ping: "pong",
+            service: "communications-manager",
+            version: version_info
+        })
+        .expect("Content-Type", /json/, done)
+});
+
+it("responds to /_status", (done) => {
+    request(server)
+        .get("/_status")
+        .expect(200, {
+            status: "pass",
+            ping: "pong",
+            service: "communications-manager",
+            version: version_info
+        })
+        .expect("Content-Type", /json/, done);
+  });
+
+  it("responds to /health", (done) => {
+    request(server)
+        .get("/health")
+        .expect(200, {
+            status: "pass",
+            ping: "pong",
+            service: "communications-manager",
+            version: version_info
+        })
+        .expect("Content-Type", /json/, done);
+  });
+})
