+++ sandbox/handlers/status.js
@@ -1,0 +1,14 @@
+async function status(req, res, next) {
+  res.json({
+      status: "pass",
+      ping: "pong",
+      service: req.app.locals.app_name,
+      version: req.app.locals.version_info
+  });
+  res.end();
+  next();
+}
+
+module.exports = {
+  status
+}
